memory/memory_management/pyproject.toml
[project]
name = "memory-management"
version = "0.1.0"
description = "Memory Management pattern for AI agents - conversation history and long-term memory"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "numpy>=1.24.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]
